Move account card to separate component
